Allow removing metrics and use normal recommendations render
